requirements.txt
streamlit
openai
python-dotenv
python-docx
langdetect
num2words